requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==3.0.6  # https://github.com/un33k/python-slugify
Pillow==6.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.4  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
flower==0.9.3  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==2.2.6  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.40.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-imagekit==4.0.2  # for resizing images
django-ckeditor==5.8.0  # ckeditor (rich text)
django-elasticsearch-dsl==7.1.1  # for searching

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
